Refactor scrolling into shared mixin
